v2.0

+++ lib/utils/loss.py
@@ -6,7 +6,7 @@ def softmax_cross_entropy_loss(input, target):
     target = target.long()
 
     e_input = torch.exp(input)
-    inds = torch.arange(len(target))
+    inds = torch.arange(len(target)).long()
     loss = -torch.log(e_input[inds, target] / torch.sum(e_input, dim=1))
     loss = loss.mean()
